Account service tests and fixes

+++ dtu_pay/account_service/src/main/java/dtu/Controllers/CustomerController.java
@@ -1,14 +1,48 @@
 package dtu.Controllers;
 
 import java.util.UUID;
+
+import org.jboss.logging.Logger;
+
+import dtu.MessagingUtils.Event;
+import dtu.MessagingUtils.MessageQueue;
 import dtu.Models.Customer;
 import dtu.Models.Database;
+import jakarta.ws.rs.NotFoundException;
 
 public class CustomerController {
     private final Database db = Database.getInstance();
+    MessageQueue queue;
+
+    private String BANKACCOUNT_CUSTOMER_REQ_RK = "payments.customerbankaccount.request";
+    private String BANKACCOUNT_CUSTOMER_RES_RK = "payments.customerbankaccount.response";
+    private String DELETE_CUSTOMER_RK = "accounts.customer.deleted";
+
+    private static final Logger LOG = Logger.getLogger(CustomerController.class);
+
+    public CustomerController(MessageQueue q) {
+        queue = q;
+
+        queue.addHandler(BANKACCOUNT_CUSTOMER_REQ_RK, e -> {
+            LOG.info("RabbitConsumer received message");
+            String accountId = e.getArgument(0, String.class);
+            String corrId = e.getArgument(1, String.class);
+
+            String bankAccountId = null;
+            if (db.hasCustomer(accountId)) {
+                bankAccountId = db.getCustomer(accountId).bankAccountUuid();
+            }
+
+            queue.publish(new Event(BANKACCOUNT_CUSTOMER_RES_RK, new Object[] { bankAccountId, corrId } ));
+		});
+    }
 
     public Customer getCustomer(String customerId) {
-        return db.getCustomer(customerId);
+        Customer customer = db.getCustomer(customerId);
+        if (customer == null) {
+            throw new NotFoundException("merchant not found");
+        }
+        return customer;
     }
 
     public Customer registerCustomer(Customer customer) {
@@ -19,7 +53,12 @@ public class CustomerController {
     }
     
     public void deleteCustomer(String id) {
+        if (!db.hasCustomer(id)) {
+            throw new NotFoundException("Customer not found");
+        }
+        
         db.deleteCustomer(id);
+        queue.publish(new Event(DELETE_CUSTOMER_RK, new Object[] { id }));
     }
 
     public boolean hasCustomer(String id) {
